Added a Suspended Transactions button
Added a suspended transactions button which is always visible. Redid the suspend button and it's visible based on if there is an item in the grid. When you click it, it pops up a menu to type in a name for the transaction.

Also added some minor logic (more of a placeholder at this point) to allow you to suspend a transaction. It uses a static ArrayList on the frmSuspendedTransactions class to populate the list and grid.