Fixed the add button for suspended transactions
The add button now will add the transaction to the main view.

Fixed a bug with closing the naming suspended transaction pop up window.

Added the ability to use the enter key when naming a transaction to be suspended

Fixed it so when a transaction is suspended it will hide the Suspend Transaction button.

Added some code so deleting a suspended transaction is possible. It is still buggy and will error a lot.